Update Dashboard User

+++ database/migrations/2026_06_24_152809_add_jenis_to_payments_table.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('payments', function (Blueprint $table) {
-            $table->string('jenis')->default('listrik')->after('id');
-        });
+        if (!Schema::hasColumn('payments', 'jenis')) {
+            Schema::table('payments', function (Blueprint $table) {
+                $table->string('jenis')->default('listrik')->after('id');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('payments', function (Blueprint $table) {
-            $table->dropColumn('jenis');
-        });
+        if (Schema::hasColumn('payments', 'jenis')) {
+            Schema::table('payments', function (Blueprint $table) {
+                $table->dropColumn('jenis');
+            });
+        }
     }
 };
